@@ -1,16 +1,16 @@
 module github.com/teslamotors/fleet-telemetry
 
-go 1.24
+go 1.25
 
 require (
-	cloud.google.com/go/pubsub v1.30.0
+	cloud.google.com/go/pubsub v1.50.1
 	github.com/airbrake/gobrake/v5 v5.6.1
 	github.com/aws/aws-sdk-go v1.44.278
 	github.com/beefsack/go-rate v0.0.0-20220214233405-116f4ca011a0
 	github.com/confluentinc/confluent-kafka-go/v2 v2.3.0
 	github.com/eclipse/paho.mqtt.golang v1.5.0
 	github.com/google/flatbuffers v23.3.3+incompatible
-	github.com/google/uuid v1.3.0
+	github.com/google/uuid v1.6.0
 	github.com/gorilla/websocket v1.5.3
 	github.com/mattn/go-colorable v0.1.13
 	github.com/onsi/ginkgo/v2 v2.22.2
