@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
-// 	protoc        v5.28.3
+// 	protoc-gen-go v1.36.0
+// 	protoc        v6.33.0
 // source: protos/vehicle_alert.proto
 
 package protos
@@ -76,22 +76,19 @@ func (Audience) EnumDescriptor() ([]byte, []int) {
 
 // VehicleAlerts is a collection of recent alerts.
 type VehicleAlerts struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Alerts        []*VehicleAlert        `protobuf:"bytes,1,rep,name=alerts,proto3" json:"alerts,omitempty"`
+	CreatedAt     *timestamppb.Timestamp `protobuf:"bytes,2,opt,name=created_at,json=createdAt,proto3" json:"created_at,omitempty"`
+	Vin           string                 `protobuf:"bytes,3,opt,name=vin,proto3" json:"vin,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Alerts    []*VehicleAlert        `protobuf:"bytes,1,rep,name=alerts,proto3" json:"alerts,omitempty"`
-	CreatedAt *timestamppb.Timestamp `protobuf:"bytes,2,opt,name=created_at,json=createdAt,proto3" json:"created_at,omitempty"`
-	Vin       string                 `protobuf:"bytes,3,opt,name=vin,proto3" json:"vin,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *VehicleAlerts) Reset() {
 	*x = VehicleAlerts{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_protos_vehicle_alert_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_protos_vehicle_alert_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *VehicleAlerts) String() string {
@@ -102,7 +99,7 @@ func (*VehicleAlerts) ProtoMessage() {}
 
 func (x *VehicleAlerts) ProtoReflect() protoreflect.Message {
 	mi := &file_protos_vehicle_alert_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -140,23 +137,20 @@ func (x *VehicleAlerts) GetVin() string {
 
 // VehicleAlert is a single alert and is active if ended_at is nil.
 type VehicleAlert struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Name          string                 `protobuf:"bytes,1,opt,name=name,proto3" json:"name,omitempty"`
+	Audiences     []Audience             `protobuf:"varint,2,rep,packed,name=audiences,proto3,enum=telemetry.vehicle_alerts.Audience" json:"audiences,omitempty"`
+	StartedAt     *timestamppb.Timestamp `protobuf:"bytes,3,opt,name=started_at,json=startedAt,proto3" json:"started_at,omitempty"`
+	EndedAt       *timestamppb.Timestamp `protobuf:"bytes,4,opt,name=ended_at,json=endedAt,proto3" json:"ended_at,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Name      string                 `protobuf:"bytes,1,opt,name=name,proto3" json:"name,omitempty"`
-	Audiences []Audience             `protobuf:"varint,2,rep,packed,name=audiences,proto3,enum=telemetry.vehicle_alerts.Audience" json:"audiences,omitempty"`
-	StartedAt *timestamppb.Timestamp `protobuf:"bytes,3,opt,name=started_at,json=startedAt,proto3" json:"started_at,omitempty"`
-	EndedAt   *timestamppb.Timestamp `protobuf:"bytes,4,opt,name=ended_at,json=endedAt,proto3" json:"ended_at,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *VehicleAlert) Reset() {
 	*x = VehicleAlert{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_protos_vehicle_alert_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_protos_vehicle_alert_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *VehicleAlert) String() string {
@@ -167,7 +161,7 @@ func (*VehicleAlert) ProtoMessage() {}
 
 func (x *VehicleAlert) ProtoReflect() protoreflect.Message {
 	mi := &file_protos_vehicle_alert_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -266,7 +260,7 @@ func file_protos_vehicle_alert_proto_rawDescGZIP() []byte {
 
 var file_protos_vehicle_alert_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_protos_vehicle_alert_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
-var file_protos_vehicle_alert_proto_goTypes = []interface{}{
+var file_protos_vehicle_alert_proto_goTypes = []any{
 	(Audience)(0),                 // 0: telemetry.vehicle_alerts.Audience
 	(*VehicleAlerts)(nil),         // 1: telemetry.vehicle_alerts.VehicleAlerts
 	(*VehicleAlert)(nil),          // 2: telemetry.vehicle_alerts.VehicleAlert
@@ -289,32 +283,6 @@ func init() { file_protos_vehicle_alert_proto_init() }
 func file_protos_vehicle_alert_proto_init() {
 	if File_protos_vehicle_alert_proto != nil {
 		return
-	}
-	if !protoimpl.UnsafeEnabled {
-		file_protos_vehicle_alert_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*VehicleAlerts); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_protos_vehicle_alert_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*VehicleAlert); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
