@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
-// 	protoc        v5.28.3
+// 	protoc-gen-go v1.36.0
+// 	protoc        v6.33.0
 // source: protos/vehicle_connectivity.proto
 
 package protos
@@ -73,24 +73,21 @@ func (ConnectivityEvent) EnumDescriptor() ([]byte, []int) {
 
 // VehicleConnectivity represents connection status change for the vehicle
 type VehicleConnectivity struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state            protoimpl.MessageState `protogen:"open.v1"`
 	Vin              string                 `protobuf:"bytes,1,opt,name=vin,proto3" json:"vin,omitempty"`
 	ConnectionId     string                 `protobuf:"bytes,2,opt,name=connection_id,json=connectionId,proto3" json:"connection_id,omitempty"`
 	Status           ConnectivityEvent      `protobuf:"varint,3,opt,name=status,proto3,enum=telemetry.vehicle_connectivity.ConnectivityEvent" json:"status,omitempty"`
 	CreatedAt        *timestamppb.Timestamp `protobuf:"bytes,4,opt,name=created_at,json=createdAt,proto3" json:"created_at,omitempty"`
 	NetworkInterface string                 `protobuf:"bytes,5,opt,name=network_interface,json=networkInterface,proto3" json:"network_interface,omitempty"`
+	unknownFields    protoimpl.UnknownFields
+	sizeCache        protoimpl.SizeCache
 }
 
 func (x *VehicleConnectivity) Reset() {
 	*x = VehicleConnectivity{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_protos_vehicle_connectivity_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_protos_vehicle_connectivity_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *VehicleConnectivity) String() string {
@@ -101,7 +98,7 @@ func (*VehicleConnectivity) ProtoMessage() {}
 
 func (x *VehicleConnectivity) ProtoReflect() protoreflect.Message {
 	mi := &file_protos_vehicle_connectivity_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -201,7 +198,7 @@ func file_protos_vehicle_connectivity_proto_rawDescGZIP() []byte {
 
 var file_protos_vehicle_connectivity_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_protos_vehicle_connectivity_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
-var file_protos_vehicle_connectivity_proto_goTypes = []interface{}{
+var file_protos_vehicle_connectivity_proto_goTypes = []any{
 	(ConnectivityEvent)(0),        // 0: telemetry.vehicle_connectivity.ConnectivityEvent
 	(*VehicleConnectivity)(nil),   // 1: telemetry.vehicle_connectivity.VehicleConnectivity
 	(*timestamppb.Timestamp)(nil), // 2: google.protobuf.Timestamp
@@ -220,20 +217,6 @@ func init() { file_protos_vehicle_connectivity_proto_init() }
 func file_protos_vehicle_connectivity_proto_init() {
 	if File_protos_vehicle_connectivity_proto != nil {
 		return
-	}
-	if !protoimpl.UnsafeEnabled {
-		file_protos_vehicle_connectivity_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*VehicleConnectivity); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
