@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
-// 	protoc        v5.28.3
+// 	protoc-gen-go v1.36.0
+// 	protoc        v6.33.0
 // source: protos/vehicle_data.proto
 
 package protos
@@ -3251,21 +3251,18 @@ func (MediaStatus) EnumDescriptor() ([]byte, []int) {
 
 // LocationValue is a Datum value type
 type LocationValue struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Latitude      float64                `protobuf:"fixed64,1,opt,name=latitude,proto3" json:"latitude,omitempty"`
+	Longitude     float64                `protobuf:"fixed64,2,opt,name=longitude,proto3" json:"longitude,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Latitude  float64 `protobuf:"fixed64,1,opt,name=latitude,proto3" json:"latitude,omitempty"`
-	Longitude float64 `protobuf:"fixed64,2,opt,name=longitude,proto3" json:"longitude,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *LocationValue) Reset() {
 	*x = LocationValue{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_protos_vehicle_data_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_protos_vehicle_data_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *LocationValue) String() string {
@@ -3276,7 +3273,7 @@ func (*LocationValue) ProtoMessage() {}
 
 func (x *LocationValue) ProtoReflect() protoreflect.Message {
 	mi := &file_protos_vehicle_data_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -3306,25 +3303,22 @@ func (x *LocationValue) GetLongitude() float64 {
 }
 
 type Doors struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	DriverFront    bool `protobuf:"varint,1,opt,name=DriverFront,proto3" json:"DriverFront,omitempty"`
-	DriverRear     bool `protobuf:"varint,2,opt,name=DriverRear,proto3" json:"DriverRear,omitempty"`
-	PassengerFront bool `protobuf:"varint,3,opt,name=PassengerFront,proto3" json:"PassengerFront,omitempty"`
-	PassengerRear  bool `protobuf:"varint,4,opt,name=PassengerRear,proto3" json:"PassengerRear,omitempty"`
-	TrunkFront     bool `protobuf:"varint,5,opt,name=TrunkFront,proto3" json:"TrunkFront,omitempty"`
-	TrunkRear      bool `protobuf:"varint,6,opt,name=TrunkRear,proto3" json:"TrunkRear,omitempty"`
+	state          protoimpl.MessageState `protogen:"open.v1"`
+	DriverFront    bool                   `protobuf:"varint,1,opt,name=DriverFront,proto3" json:"DriverFront,omitempty"`
+	DriverRear     bool                   `protobuf:"varint,2,opt,name=DriverRear,proto3" json:"DriverRear,omitempty"`
+	PassengerFront bool                   `protobuf:"varint,3,opt,name=PassengerFront,proto3" json:"PassengerFront,omitempty"`
+	PassengerRear  bool                   `protobuf:"varint,4,opt,name=PassengerRear,proto3" json:"PassengerRear,omitempty"`
+	TrunkFront     bool                   `protobuf:"varint,5,opt,name=TrunkFront,proto3" json:"TrunkFront,omitempty"`
+	TrunkRear      bool                   `protobuf:"varint,6,opt,name=TrunkRear,proto3" json:"TrunkRear,omitempty"`
+	unknownFields  protoimpl.UnknownFields
+	sizeCache      protoimpl.SizeCache
 }
 
 func (x *Doors) Reset() {
 	*x = Doors{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_protos_vehicle_data_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_protos_vehicle_data_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Doors) String() string {
@@ -3335,7 +3329,7 @@ func (*Doors) ProtoMessage() {}
 
 func (x *Doors) ProtoReflect() protoreflect.Message {
 	mi := &file_protos_vehicle_data_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -3393,29 +3387,26 @@ func (x *Doors) GetTrunkRear() bool {
 }
 
 type TireLocation struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	FrontLeft             bool `protobuf:"varint,1,opt,name=front_left,json=frontLeft,proto3" json:"front_left,omitempty"`
-	FrontRight            bool `protobuf:"varint,2,opt,name=front_right,json=frontRight,proto3" json:"front_right,omitempty"`
-	RearLeft              bool `protobuf:"varint,3,opt,name=rear_left,json=rearLeft,proto3" json:"rear_left,omitempty"`
-	RearRight             bool `protobuf:"varint,4,opt,name=rear_right,json=rearRight,proto3" json:"rear_right,omitempty"`
-	SemiMiddleAxleLeft_2  bool `protobuf:"varint,5,opt,name=semi_middle_axle_left_2,json=semiMiddleAxleLeft2,proto3" json:"semi_middle_axle_left_2,omitempty"`
-	SemiMiddleAxleRight_2 bool `protobuf:"varint,6,opt,name=semi_middle_axle_right_2,json=semiMiddleAxleRight2,proto3" json:"semi_middle_axle_right_2,omitempty"`
-	SemiRearAxleLeft      bool `protobuf:"varint,7,opt,name=semi_rear_axle_left,json=semiRearAxleLeft,proto3" json:"semi_rear_axle_left,omitempty"`
-	SemiRearAxleRight     bool `protobuf:"varint,8,opt,name=semi_rear_axle_right,json=semiRearAxleRight,proto3" json:"semi_rear_axle_right,omitempty"`
-	SemiRearAxleLeft_2    bool `protobuf:"varint,9,opt,name=semi_rear_axle_left_2,json=semiRearAxleLeft2,proto3" json:"semi_rear_axle_left_2,omitempty"`
-	SemiRearAxleRight_2   bool `protobuf:"varint,10,opt,name=semi_rear_axle_right_2,json=semiRearAxleRight2,proto3" json:"semi_rear_axle_right_2,omitempty"`
+	state                 protoimpl.MessageState `protogen:"open.v1"`
+	FrontLeft             bool                   `protobuf:"varint,1,opt,name=front_left,json=frontLeft,proto3" json:"front_left,omitempty"`
+	FrontRight            bool                   `protobuf:"varint,2,opt,name=front_right,json=frontRight,proto3" json:"front_right,omitempty"`
+	RearLeft              bool                   `protobuf:"varint,3,opt,name=rear_left,json=rearLeft,proto3" json:"rear_left,omitempty"`
+	RearRight             bool                   `protobuf:"varint,4,opt,name=rear_right,json=rearRight,proto3" json:"rear_right,omitempty"`
+	SemiMiddleAxleLeft_2  bool                   `protobuf:"varint,5,opt,name=semi_middle_axle_left_2,json=semiMiddleAxleLeft2,proto3" json:"semi_middle_axle_left_2,omitempty"`
+	SemiMiddleAxleRight_2 bool                   `protobuf:"varint,6,opt,name=semi_middle_axle_right_2,json=semiMiddleAxleRight2,proto3" json:"semi_middle_axle_right_2,omitempty"`
+	SemiRearAxleLeft      bool                   `protobuf:"varint,7,opt,name=semi_rear_axle_left,json=semiRearAxleLeft,proto3" json:"semi_rear_axle_left,omitempty"`
+	SemiRearAxleRight     bool                   `protobuf:"varint,8,opt,name=semi_rear_axle_right,json=semiRearAxleRight,proto3" json:"semi_rear_axle_right,omitempty"`
+	SemiRearAxleLeft_2    bool                   `protobuf:"varint,9,opt,name=semi_rear_axle_left_2,json=semiRearAxleLeft2,proto3" json:"semi_rear_axle_left_2,omitempty"`
+	SemiRearAxleRight_2   bool                   `protobuf:"varint,10,opt,name=semi_rear_axle_right_2,json=semiRearAxleRight2,proto3" json:"semi_rear_axle_right_2,omitempty"`
+	unknownFields         protoimpl.UnknownFields
+	sizeCache             protoimpl.SizeCache
 }
 
 func (x *TireLocation) Reset() {
 	*x = TireLocation{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_protos_vehicle_data_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_protos_vehicle_data_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TireLocation) String() string {
@@ -3426,7 +3417,7 @@ func (*TireLocation) ProtoMessage() {}
 
 func (x *TireLocation) ProtoReflect() protoreflect.Message {
 	mi := &file_protos_vehicle_data_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -3512,22 +3503,19 @@ func (x *TireLocation) GetSemiRearAxleRight_2() bool {
 }
 
 type Time struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Hour          int32                  `protobuf:"varint,1,opt,name=hour,proto3" json:"hour,omitempty"`
+	Minute        int32                  `protobuf:"varint,2,opt,name=minute,proto3" json:"minute,omitempty"`
+	Second        int32                  `protobuf:"varint,3,opt,name=second,proto3" json:"second,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Hour   int32 `protobuf:"varint,1,opt,name=hour,proto3" json:"hour,omitempty"`
-	Minute int32 `protobuf:"varint,2,opt,name=minute,proto3" json:"minute,omitempty"`
-	Second int32 `protobuf:"varint,3,opt,name=second,proto3" json:"second,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Time) Reset() {
 	*x = Time{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_protos_vehicle_data_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_protos_vehicle_data_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Time) String() string {
@@ -3538,7 +3526,7 @@ func (*Time) ProtoMessage() {}
 
 func (x *Time) ProtoReflect() protoreflect.Message {
 	mi := &file_protos_vehicle_data_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -3578,11 +3566,8 @@ func (x *Time) GetSecond() int32 {
 // Note: Fields may have their types updated with different software and vehicle
 // versions to optimize for precision or space
 type Value struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	// Types that are assignable to Value:
+	state protoimpl.MessageState `protogen:"open.v1"`
+	// Types that are valid to be assigned to Value:
 	//
 	//	*Value_StringValue
 	//	*Value_IntValue
@@ -3637,16 +3622,16 @@ type Value struct {
 	//	*Value_TurnSignalStateValue
 	//	*Value_MediaStatusValue
 	//	*Value_SunroofInstalledStateValue
-	Value isValue_Value `protobuf_oneof:"value"`
+	Value         isValue_Value `protobuf_oneof:"value"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Value) Reset() {
 	*x = Value{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_protos_vehicle_data_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_protos_vehicle_data_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Value) String() string {
@@ -3657,7 +3642,7 @@ func (*Value) ProtoMessage() {}
 
 func (x *Value) ProtoReflect() protoreflect.Message {
 	mi := &file_protos_vehicle_data_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -3672,380 +3657,486 @@ func (*Value) Descriptor() ([]byte, []int) {
 	return file_protos_vehicle_data_proto_rawDescGZIP(), []int{4}
 }
 
-func (m *Value) GetValue() isValue_Value {
-	if m != nil {
-		return m.Value
+func (x *Value) GetValue() isValue_Value {
+	if x != nil {
+		return x.Value
 	}
 	return nil
 }
 
 func (x *Value) GetStringValue() string {
-	if x, ok := x.GetValue().(*Value_StringValue); ok {
-		return x.StringValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_StringValue); ok {
+			return x.StringValue
+		}
 	}
 	return ""
 }
 
 func (x *Value) GetIntValue() int32 {
-	if x, ok := x.GetValue().(*Value_IntValue); ok {
-		return x.IntValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_IntValue); ok {
+			return x.IntValue
+		}
 	}
 	return 0
 }
 
 func (x *Value) GetLongValue() int64 {
-	if x, ok := x.GetValue().(*Value_LongValue); ok {
-		return x.LongValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_LongValue); ok {
+			return x.LongValue
+		}
 	}
 	return 0
 }
 
 func (x *Value) GetFloatValue() float32 {
-	if x, ok := x.GetValue().(*Value_FloatValue); ok {
-		return x.FloatValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_FloatValue); ok {
+			return x.FloatValue
+		}
 	}
 	return 0
 }
 
 func (x *Value) GetDoubleValue() float64 {
-	if x, ok := x.GetValue().(*Value_DoubleValue); ok {
-		return x.DoubleValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_DoubleValue); ok {
+			return x.DoubleValue
+		}
 	}
 	return 0
 }
 
 func (x *Value) GetBooleanValue() bool {
-	if x, ok := x.GetValue().(*Value_BooleanValue); ok {
-		return x.BooleanValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_BooleanValue); ok {
+			return x.BooleanValue
+		}
 	}
 	return false
 }
 
 func (x *Value) GetLocationValue() *LocationValue {
-	if x, ok := x.GetValue().(*Value_LocationValue); ok {
-		return x.LocationValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_LocationValue); ok {
+			return x.LocationValue
+		}
 	}
 	return nil
 }
 
 func (x *Value) GetChargingValue() ChargingState {
-	if x, ok := x.GetValue().(*Value_ChargingValue); ok {
-		return x.ChargingValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_ChargingValue); ok {
+			return x.ChargingValue
+		}
 	}
 	return ChargingState_ChargeStateUnknown
 }
 
 func (x *Value) GetShiftStateValue() ShiftState {
-	if x, ok := x.GetValue().(*Value_ShiftStateValue); ok {
-		return x.ShiftStateValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_ShiftStateValue); ok {
+			return x.ShiftStateValue
+		}
 	}
 	return ShiftState_ShiftStateUnknown
 }
 
 func (x *Value) GetInvalid() bool {
-	if x, ok := x.GetValue().(*Value_Invalid); ok {
-		return x.Invalid
+	if x != nil {
+		if x, ok := x.Value.(*Value_Invalid); ok {
+			return x.Invalid
+		}
 	}
 	return false
 }
 
 func (x *Value) GetLaneAssistLevelValue() LaneAssistLevel {
-	if x, ok := x.GetValue().(*Value_LaneAssistLevelValue); ok {
-		return x.LaneAssistLevelValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_LaneAssistLevelValue); ok {
+			return x.LaneAssistLevelValue
+		}
 	}
 	return LaneAssistLevel_LaneAssistLevelUnknown
 }
 
 func (x *Value) GetScheduledChargingModeValue() ScheduledChargingModeValue {
-	if x, ok := x.GetValue().(*Value_ScheduledChargingModeValue); ok {
-		return x.ScheduledChargingModeValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_ScheduledChargingModeValue); ok {
+			return x.ScheduledChargingModeValue
+		}
 	}
 	return ScheduledChargingModeValue_ScheduledChargingModeUnknown
 }
 
 func (x *Value) GetSentryModeStateValue() SentryModeState {
-	if x, ok := x.GetValue().(*Value_SentryModeStateValue); ok {
-		return x.SentryModeStateValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_SentryModeStateValue); ok {
+			return x.SentryModeStateValue
+		}
 	}
 	return SentryModeState_SentryModeStateUnknown
 }
 
 func (x *Value) GetSpeedAssistLevelValue() SpeedAssistLevel {
-	if x, ok := x.GetValue().(*Value_SpeedAssistLevelValue); ok {
-		return x.SpeedAssistLevelValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_SpeedAssistLevelValue); ok {
+			return x.SpeedAssistLevelValue
+		}
 	}
 	return SpeedAssistLevel_SpeedAssistLevelUnknown
 }
 
 func (x *Value) GetBmsStateValue() BMSStateValue {
-	if x, ok := x.GetValue().(*Value_BmsStateValue); ok {
-		return x.BmsStateValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_BmsStateValue); ok {
+			return x.BmsStateValue
+		}
 	}
 	return BMSStateValue_BMSStateUnknown
 }
 
 func (x *Value) GetBuckleStatusValue() BuckleStatus {
-	if x, ok := x.GetValue().(*Value_BuckleStatusValue); ok {
-		return x.BuckleStatusValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_BuckleStatusValue); ok {
+			return x.BuckleStatusValue
+		}
 	}
 	return BuckleStatus_BuckleStatusUnknown
 }
 
 func (x *Value) GetCarTypeValue() CarTypeValue {
-	if x, ok := x.GetValue().(*Value_CarTypeValue); ok {
-		return x.CarTypeValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_CarTypeValue); ok {
+			return x.CarTypeValue
+		}
 	}
 	return CarTypeValue_CarTypeUnknown
 }
 
 func (x *Value) GetChargePortValue() ChargePortValue {
-	if x, ok := x.GetValue().(*Value_ChargePortValue); ok {
-		return x.ChargePortValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_ChargePortValue); ok {
+			return x.ChargePortValue
+		}
 	}
 	return ChargePortValue_ChargePortUnknown
 }
 
 func (x *Value) GetChargePortLatchValue() ChargePortLatchValue {
-	if x, ok := x.GetValue().(*Value_ChargePortLatchValue); ok {
-		return x.ChargePortLatchValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_ChargePortLatchValue); ok {
+			return x.ChargePortLatchValue
+		}
 	}
 	return ChargePortLatchValue_ChargePortLatchUnknown
 }
 
 func (x *Value) GetDoorValue() *Doors {
-	if x, ok := x.GetValue().(*Value_DoorValue); ok {
-		return x.DoorValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_DoorValue); ok {
+			return x.DoorValue
+		}
 	}
 	return nil
 }
 
 func (x *Value) GetDriveInverterStateValue() DriveInverterState {
-	if x, ok := x.GetValue().(*Value_DriveInverterStateValue); ok {
-		return x.DriveInverterStateValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_DriveInverterStateValue); ok {
+			return x.DriveInverterStateValue
+		}
 	}
 	return DriveInverterState_DriveInverterStateUnknown
 }
 
 func (x *Value) GetHvilStatusValue() HvilStatus {
-	if x, ok := x.GetValue().(*Value_HvilStatusValue); ok {
-		return x.HvilStatusValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_HvilStatusValue); ok {
+			return x.HvilStatusValue
+		}
 	}
 	return HvilStatus_HvilStatusUnknown
 }
 
 func (x *Value) GetWindowStateValue() WindowState {
-	if x, ok := x.GetValue().(*Value_WindowStateValue); ok {
-		return x.WindowStateValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_WindowStateValue); ok {
+			return x.WindowStateValue
+		}
 	}
 	return WindowState_WindowStateUnknown
 }
 
 func (x *Value) GetSeatFoldPositionValue() SeatFoldPosition {
-	if x, ok := x.GetValue().(*Value_SeatFoldPositionValue); ok {
-		return x.SeatFoldPositionValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_SeatFoldPositionValue); ok {
+			return x.SeatFoldPositionValue
+		}
 	}
 	return SeatFoldPosition_SeatFoldPositionUnknown
 }
 
 func (x *Value) GetTractorAirStatusValue() TractorAirStatus {
-	if x, ok := x.GetValue().(*Value_TractorAirStatusValue); ok {
-		return x.TractorAirStatusValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_TractorAirStatusValue); ok {
+			return x.TractorAirStatusValue
+		}
 	}
 	return TractorAirStatus_TractorAirStatusUnknown
 }
 
 func (x *Value) GetFollowDistanceValue() FollowDistance {
-	if x, ok := x.GetValue().(*Value_FollowDistanceValue); ok {
-		return x.FollowDistanceValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_FollowDistanceValue); ok {
+			return x.FollowDistanceValue
+		}
 	}
 	return FollowDistance_FollowDistanceUnknown
 }
 
 func (x *Value) GetForwardCollisionSensitivityValue() ForwardCollisionSensitivity {
-	if x, ok := x.GetValue().(*Value_ForwardCollisionSensitivityValue); ok {
-		return x.ForwardCollisionSensitivityValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_ForwardCollisionSensitivityValue); ok {
+			return x.ForwardCollisionSensitivityValue
+		}
 	}
 	return ForwardCollisionSensitivity_ForwardCollisionSensitivityUnknown
 }
 
 func (x *Value) GetGuestModeMobileAccessValue() GuestModeMobileAccess {
-	if x, ok := x.GetValue().(*Value_GuestModeMobileAccessValue); ok {
-		return x.GuestModeMobileAccessValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_GuestModeMobileAccessValue); ok {
+			return x.GuestModeMobileAccessValue
+		}
 	}
 	return GuestModeMobileAccess_GuestModeMobileAccessUnknown
 }
 
 func (x *Value) GetTrailerAirStatusValue() TrailerAirStatus {
-	if x, ok := x.GetValue().(*Value_TrailerAirStatusValue); ok {
-		return x.TrailerAirStatusValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_TrailerAirStatusValue); ok {
+			return x.TrailerAirStatusValue
+		}
 	}
 	return TrailerAirStatus_TrailerAirStatusUnknown
 }
 
 func (x *Value) GetTimeValue() *Time {
-	if x, ok := x.GetValue().(*Value_TimeValue); ok {
-		return x.TimeValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_TimeValue); ok {
+			return x.TimeValue
+		}
 	}
 	return nil
 }
 
 func (x *Value) GetDetailedChargeStateValue() DetailedChargeStateValue {
-	if x, ok := x.GetValue().(*Value_DetailedChargeStateValue); ok {
-		return x.DetailedChargeStateValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_DetailedChargeStateValue); ok {
+			return x.DetailedChargeStateValue
+		}
 	}
 	return DetailedChargeStateValue_DetailedChargeStateUnknown
 }
 
 func (x *Value) GetHvacAutoModeValue() HvacAutoModeState {
-	if x, ok := x.GetValue().(*Value_HvacAutoModeValue); ok {
-		return x.HvacAutoModeValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_HvacAutoModeValue); ok {
+			return x.HvacAutoModeValue
+		}
 	}
 	return HvacAutoModeState_HvacAutoModeStateUnknown
 }
 
 func (x *Value) GetCabinOverheatProtectionModeValue() CabinOverheatProtectionModeState {
-	if x, ok := x.GetValue().(*Value_CabinOverheatProtectionModeValue); ok {
-		return x.CabinOverheatProtectionModeValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_CabinOverheatProtectionModeValue); ok {
+			return x.CabinOverheatProtectionModeValue
+		}
 	}
 	return CabinOverheatProtectionModeState_CabinOverheatProtectionModeStateUnknown
 }
 
 func (x *Value) GetCabinOverheatProtectionTemperatureLimitValue() ClimateOverheatProtectionTempLimit {
-	if x, ok := x.GetValue().(*Value_CabinOverheatProtectionTemperatureLimitValue); ok {
-		return x.CabinOverheatProtectionTemperatureLimitValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_CabinOverheatProtectionTemperatureLimitValue); ok {
+			return x.CabinOverheatProtectionTemperatureLimitValue
+		}
 	}
 	return ClimateOverheatProtectionTempLimit_ClimateOverheatProtectionTempLimitUnknown
 }
 
 func (x *Value) GetDefrostModeValue() DefrostModeState {
-	if x, ok := x.GetValue().(*Value_DefrostModeValue); ok {
-		return x.DefrostModeValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_DefrostModeValue); ok {
+			return x.DefrostModeValue
+		}
 	}
 	return DefrostModeState_DefrostModeStateUnknown
 }
 
 func (x *Value) GetClimateKeeperModeValue() ClimateKeeperModeState {
-	if x, ok := x.GetValue().(*Value_ClimateKeeperModeValue); ok {
-		return x.ClimateKeeperModeValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_ClimateKeeperModeValue); ok {
+			return x.ClimateKeeperModeValue
+		}
 	}
 	return ClimateKeeperModeState_ClimateKeeperModeStateUnknown
 }
 
 func (x *Value) GetHvacPowerValue() HvacPowerState {
-	if x, ok := x.GetValue().(*Value_HvacPowerValue); ok {
-		return x.HvacPowerValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_HvacPowerValue); ok {
+			return x.HvacPowerValue
+		}
 	}
 	return HvacPowerState_HvacPowerStateUnknown
 }
 
 func (x *Value) GetTireLocationValue() *TireLocation {
-	if x, ok := x.GetValue().(*Value_TireLocationValue); ok {
-		return x.TireLocationValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_TireLocationValue); ok {
+			return x.TireLocationValue
+		}
 	}
 	return nil
 }
 
 func (x *Value) GetFastChargerValue() FastCharger {
-	if x, ok := x.GetValue().(*Value_FastChargerValue); ok {
-		return x.FastChargerValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_FastChargerValue); ok {
+			return x.FastChargerValue
+		}
 	}
 	return FastCharger_FastChargerUnknown
 }
 
 func (x *Value) GetCableTypeValue() CableType {
-	if x, ok := x.GetValue().(*Value_CableTypeValue); ok {
-		return x.CableTypeValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_CableTypeValue); ok {
+			return x.CableTypeValue
+		}
 	}
 	return CableType_CableTypeUnknown
 }
 
 func (x *Value) GetTonneauTentModeValue() TonneauTentModeState {
-	if x, ok := x.GetValue().(*Value_TonneauTentModeValue); ok {
-		return x.TonneauTentModeValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_TonneauTentModeValue); ok {
+			return x.TonneauTentModeValue
+		}
 	}
 	return TonneauTentModeState_TonneauTentModeStateUnknown
 }
 
 func (x *Value) GetTonneauPositionValue() TonneauPositionState {
-	if x, ok := x.GetValue().(*Value_TonneauPositionValue); ok {
-		return x.TonneauPositionValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_TonneauPositionValue); ok {
+			return x.TonneauPositionValue
+		}
 	}
 	return TonneauPositionState_TonneauPositionStateUnknown
 }
 
 func (x *Value) GetPowershareTypeValue() PowershareTypeStatus {
-	if x, ok := x.GetValue().(*Value_PowershareTypeValue); ok {
-		return x.PowershareTypeValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_PowershareTypeValue); ok {
+			return x.PowershareTypeValue
+		}
 	}
 	return PowershareTypeStatus_PowershareTypeStatusUnknown
 }
 
 func (x *Value) GetPowershareStateValue() PowershareState {
-	if x, ok := x.GetValue().(*Value_PowershareStateValue); ok {
-		return x.PowershareStateValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_PowershareStateValue); ok {
+			return x.PowershareStateValue
+		}
 	}
 	return PowershareState_PowershareStateUnknown
 }
 
 func (x *Value) GetPowershareStopReasonValue() PowershareStopReasonStatus {
-	if x, ok := x.GetValue().(*Value_PowershareStopReasonValue); ok {
-		return x.PowershareStopReasonValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_PowershareStopReasonValue); ok {
+			return x.PowershareStopReasonValue
+		}
 	}
 	return PowershareStopReasonStatus_PowershareStopReasonStatusUnknown
 }
 
 func (x *Value) GetDisplayStateValue() DisplayState {
-	if x, ok := x.GetValue().(*Value_DisplayStateValue); ok {
-		return x.DisplayStateValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_DisplayStateValue); ok {
+			return x.DisplayStateValue
+		}
 	}
 	return DisplayState_DisplayStateUnknown
 }
 
 func (x *Value) GetDistanceUnitValue() DistanceUnit {
-	if x, ok := x.GetValue().(*Value_DistanceUnitValue); ok {
-		return x.DistanceUnitValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_DistanceUnitValue); ok {
+			return x.DistanceUnitValue
+		}
 	}
 	return DistanceUnit_DistanceUnitUnknown
 }
 
 func (x *Value) GetTemperatureUnitValue() TemperatureUnit {
-	if x, ok := x.GetValue().(*Value_TemperatureUnitValue); ok {
-		return x.TemperatureUnitValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_TemperatureUnitValue); ok {
+			return x.TemperatureUnitValue
+		}
 	}
 	return TemperatureUnit_TemperatureUnitUnknown
 }
 
 func (x *Value) GetPressureUnitValue() PressureUnit {
-	if x, ok := x.GetValue().(*Value_PressureUnitValue); ok {
-		return x.PressureUnitValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_PressureUnitValue); ok {
+			return x.PressureUnitValue
+		}
 	}
 	return PressureUnit_PressureUnitUnknown
 }
 
 func (x *Value) GetChargeUnitPreferenceValue() ChargeUnitPreference {
-	if x, ok := x.GetValue().(*Value_ChargeUnitPreferenceValue); ok {
-		return x.ChargeUnitPreferenceValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_ChargeUnitPreferenceValue); ok {
+			return x.ChargeUnitPreferenceValue
+		}
 	}
 	return ChargeUnitPreference_ChargeUnitUnknown
 }
 
 func (x *Value) GetTurnSignalStateValue() TurnSignalState {
-	if x, ok := x.GetValue().(*Value_TurnSignalStateValue); ok {
-		return x.TurnSignalStateValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_TurnSignalStateValue); ok {
+			return x.TurnSignalStateValue
+		}
 	}
 	return TurnSignalState_TurnSignalStateUnknown
 }
 
 func (x *Value) GetMediaStatusValue() MediaStatus {
-	if x, ok := x.GetValue().(*Value_MediaStatusValue); ok {
-		return x.MediaStatusValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_MediaStatusValue); ok {
+			return x.MediaStatusValue
+		}
 	}
 	return MediaStatus_MediaStatusUnknown
 }
 
 func (x *Value) GetSunroofInstalledStateValue() SunroofInstalledState {
-	if x, ok := x.GetValue().(*Value_SunroofInstalledStateValue); ok {
-		return x.SunroofInstalledStateValue
+	if x != nil {
+		if x, ok := x.Value.(*Value_SunroofInstalledStateValue); ok {
+			return x.SunroofInstalledStateValue
+		}
 	}
 	return SunroofInstalledState_SunroofInstalledStateUnknown
 }
@@ -4374,21 +4465,18 @@ func (*Value_SunroofInstalledStateValue) isValue_Value() {}
 
 // Datum represents a single field and its value
 type Datum struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Key           Field                  `protobuf:"varint,1,opt,name=key,proto3,enum=telemetry.vehicle_data.Field" json:"key,omitempty"`
+	Value         *Value                 `protobuf:"bytes,2,opt,name=value,proto3" json:"value,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Key   Field  `protobuf:"varint,1,opt,name=key,proto3,enum=telemetry.vehicle_data.Field" json:"key,omitempty"`
-	Value *Value `protobuf:"bytes,2,opt,name=value,proto3" json:"value,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Datum) Reset() {
 	*x = Datum{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_protos_vehicle_data_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_protos_vehicle_data_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Datum) String() string {
@@ -4399,7 +4487,7 @@ func (*Datum) ProtoMessage() {}
 
 func (x *Datum) ProtoReflect() protoreflect.Message {
 	mi := &file_protos_vehicle_data_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -4430,23 +4518,20 @@ func (x *Datum) GetValue() *Value {
 
 // Payload holds a collection of values, and may not have a vin
 type Payload struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Data          []*Datum               `protobuf:"bytes,1,rep,name=data,proto3" json:"data,omitempty"`
+	CreatedAt     *timestamppb.Timestamp `protobuf:"bytes,2,opt,name=created_at,json=createdAt,proto3" json:"created_at,omitempty"`
+	Vin           string                 `protobuf:"bytes,3,opt,name=vin,proto3" json:"vin,omitempty"`
+	IsResend      bool                   `protobuf:"varint,4,opt,name=is_resend,json=isResend,proto3" json:"is_resend,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Data      []*Datum               `protobuf:"bytes,1,rep,name=data,proto3" json:"data,omitempty"`
-	CreatedAt *timestamppb.Timestamp `protobuf:"bytes,2,opt,name=created_at,json=createdAt,proto3" json:"created_at,omitempty"`
-	Vin       string                 `protobuf:"bytes,3,opt,name=vin,proto3" json:"vin,omitempty"`
-	IsResend  bool                   `protobuf:"varint,4,opt,name=is_resend,json=isResend,proto3" json:"is_resend,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Payload) Reset() {
 	*x = Payload{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_protos_vehicle_data_proto_msgTypes[6]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_protos_vehicle_data_proto_msgTypes[6]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Payload) String() string {
@@ -4457,7 +4542,7 @@ func (*Payload) ProtoMessage() {}
 
 func (x *Payload) ProtoReflect() protoreflect.Message {
 	mi := &file_protos_vehicle_data_proto_msgTypes[6]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -5778,7 +5863,7 @@ func file_protos_vehicle_data_proto_rawDescGZIP() []byte {
 
 var file_protos_vehicle_data_proto_enumTypes = make([]protoimpl.EnumInfo, 43)
 var file_protos_vehicle_data_proto_msgTypes = make([]protoimpl.MessageInfo, 7)
-var file_protos_vehicle_data_proto_goTypes = []interface{}{
+var file_protos_vehicle_data_proto_goTypes = []any{
 	(Field)(0),                              // 0: telemetry.vehicle_data.Field
 	(ChargingState)(0),                      // 1: telemetry.vehicle_data.ChargingState
 	(DetailedChargeStateValue)(0),           // 2: telemetry.vehicle_data.DetailedChargeStateValue
@@ -5894,93 +5979,7 @@ func file_protos_vehicle_data_proto_init() {
 	if File_protos_vehicle_data_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_protos_vehicle_data_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*LocationValue); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_protos_vehicle_data_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Doors); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_protos_vehicle_data_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TireLocation); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_protos_vehicle_data_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Time); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_protos_vehicle_data_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Value); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_protos_vehicle_data_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Datum); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_protos_vehicle_data_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Payload); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
-	file_protos_vehicle_data_proto_msgTypes[4].OneofWrappers = []interface{}{
+	file_protos_vehicle_data_proto_msgTypes[4].OneofWrappers = []any{
 		(*Value_StringValue)(nil),
 		(*Value_IntValue)(nil),
 		(*Value_LongValue)(nil),
