@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
-// 	protoc        v5.28.3
+// 	protoc-gen-go v1.36.0
+// 	protoc        v6.33.0
 // source: protos/vehicle_error.proto
 
 package protos
@@ -23,22 +23,19 @@ const (
 
 // VehicleErrors is a collection of errors for a single vehicle.
 type VehicleErrors struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Errors        []*VehicleError        `protobuf:"bytes,1,rep,name=errors,proto3" json:"errors,omitempty"`
+	CreatedAt     *timestamppb.Timestamp `protobuf:"bytes,2,opt,name=created_at,json=createdAt,proto3" json:"created_at,omitempty"`
+	Vin           string                 `protobuf:"bytes,3,opt,name=vin,proto3" json:"vin,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Errors    []*VehicleError        `protobuf:"bytes,1,rep,name=errors,proto3" json:"errors,omitempty"`
-	CreatedAt *timestamppb.Timestamp `protobuf:"bytes,2,opt,name=created_at,json=createdAt,proto3" json:"created_at,omitempty"`
-	Vin       string                 `protobuf:"bytes,3,opt,name=vin,proto3" json:"vin,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *VehicleErrors) Reset() {
 	*x = VehicleErrors{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_protos_vehicle_error_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_protos_vehicle_error_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *VehicleErrors) String() string {
@@ -49,7 +46,7 @@ func (*VehicleErrors) ProtoMessage() {}
 
 func (x *VehicleErrors) ProtoReflect() protoreflect.Message {
 	mi := &file_protos_vehicle_error_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -87,23 +84,20 @@ func (x *VehicleErrors) GetVin() string {
 
 // VehicleError is a single error
 type VehicleError struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	CreatedAt     *timestamppb.Timestamp `protobuf:"bytes,1,opt,name=created_at,json=createdAt,proto3" json:"created_at,omitempty"`
+	Name          string                 `protobuf:"bytes,2,opt,name=name,proto3" json:"name,omitempty"`
+	Tags          map[string]string      `protobuf:"bytes,3,rep,name=tags,proto3" json:"tags,omitempty" protobuf_key:"bytes,1,opt,name=key" protobuf_val:"bytes,2,opt,name=value"`
+	Body          string                 `protobuf:"bytes,4,opt,name=body,proto3" json:"body,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	CreatedAt *timestamppb.Timestamp `protobuf:"bytes,1,opt,name=created_at,json=createdAt,proto3" json:"created_at,omitempty"`
-	Name      string                 `protobuf:"bytes,2,opt,name=name,proto3" json:"name,omitempty"`
-	Tags      map[string]string      `protobuf:"bytes,3,rep,name=tags,proto3" json:"tags,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3"`
-	Body      string                 `protobuf:"bytes,4,opt,name=body,proto3" json:"body,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *VehicleError) Reset() {
 	*x = VehicleError{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_protos_vehicle_error_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_protos_vehicle_error_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *VehicleError) String() string {
@@ -114,7 +108,7 @@ func (*VehicleError) ProtoMessage() {}
 
 func (x *VehicleError) ProtoReflect() protoreflect.Message {
 	mi := &file_protos_vehicle_error_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -209,7 +203,7 @@ func file_protos_vehicle_error_proto_rawDescGZIP() []byte {
 }
 
 var file_protos_vehicle_error_proto_msgTypes = make([]protoimpl.MessageInfo, 3)
-var file_protos_vehicle_error_proto_goTypes = []interface{}{
+var file_protos_vehicle_error_proto_goTypes = []any{
 	(*VehicleErrors)(nil),         // 0: telemetry.vehicle_error.VehicleErrors
 	(*VehicleError)(nil),          // 1: telemetry.vehicle_error.VehicleError
 	nil,                           // 2: telemetry.vehicle_error.VehicleError.TagsEntry
@@ -231,32 +225,6 @@ func init() { file_protos_vehicle_error_proto_init() }
 func file_protos_vehicle_error_proto_init() {
 	if File_protos_vehicle_error_proto != nil {
 		return
-	}
-	if !protoimpl.UnsafeEnabled {
-		file_protos_vehicle_error_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*VehicleErrors); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_protos_vehicle_error_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*VehicleError); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
