@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
-// 	protoc        v5.28.3
+// 	protoc-gen-go v1.36.0
+// 	protoc        v6.33.0
 // source: protos/vehicle_metric.proto
 
 package protos
@@ -23,22 +23,19 @@ const (
 
 // VehicleMetrics is a collection of metrics for a single vehicle.
 type VehicleMetrics struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Metrics       []*Metric              `protobuf:"bytes,1,rep,name=metrics,proto3" json:"metrics,omitempty"`
+	CreatedAt     *timestamppb.Timestamp `protobuf:"bytes,2,opt,name=created_at,json=createdAt,proto3" json:"created_at,omitempty"`
+	Vin           string                 `protobuf:"bytes,3,opt,name=vin,proto3" json:"vin,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Metrics   []*Metric              `protobuf:"bytes,1,rep,name=metrics,proto3" json:"metrics,omitempty"`
-	CreatedAt *timestamppb.Timestamp `protobuf:"bytes,2,opt,name=created_at,json=createdAt,proto3" json:"created_at,omitempty"`
-	Vin       string                 `protobuf:"bytes,3,opt,name=vin,proto3" json:"vin,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *VehicleMetrics) Reset() {
 	*x = VehicleMetrics{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_protos_vehicle_metric_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_protos_vehicle_metric_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *VehicleMetrics) String() string {
@@ -49,7 +46,7 @@ func (*VehicleMetrics) ProtoMessage() {}
 
 func (x *VehicleMetrics) ProtoReflect() protoreflect.Message {
 	mi := &file_protos_vehicle_metric_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -87,22 +84,19 @@ func (x *VehicleMetrics) GetVin() string {
 
 // Metric is a single metric value with a name, tags, and a value.
 type Metric struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Name          string                 `protobuf:"bytes,1,opt,name=name,proto3" json:"name,omitempty"`
+	Tags          map[string]string      `protobuf:"bytes,2,rep,name=tags,proto3" json:"tags,omitempty" protobuf_key:"bytes,1,opt,name=key" protobuf_val:"bytes,2,opt,name=value"`
+	Value         float64                `protobuf:"fixed64,3,opt,name=value,proto3" json:"value,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Name  string            `protobuf:"bytes,1,opt,name=name,proto3" json:"name,omitempty"`
-	Tags  map[string]string `protobuf:"bytes,2,rep,name=tags,proto3" json:"tags,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3"`
-	Value float64           `protobuf:"fixed64,3,opt,name=value,proto3" json:"value,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Metric) Reset() {
 	*x = Metric{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_protos_vehicle_metric_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_protos_vehicle_metric_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Metric) String() string {
@@ -113,7 +107,7 @@ func (*Metric) ProtoMessage() {}
 
 func (x *Metric) ProtoReflect() protoreflect.Message {
 	mi := &file_protos_vehicle_metric_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -197,7 +191,7 @@ func file_protos_vehicle_metric_proto_rawDescGZIP() []byte {
 }
 
 var file_protos_vehicle_metric_proto_msgTypes = make([]protoimpl.MessageInfo, 3)
-var file_protos_vehicle_metric_proto_goTypes = []interface{}{
+var file_protos_vehicle_metric_proto_goTypes = []any{
 	(*VehicleMetrics)(nil),        // 0: telemetry.vehicle_metrics.VehicleMetrics
 	(*Metric)(nil),                // 1: telemetry.vehicle_metrics.Metric
 	nil,                           // 2: telemetry.vehicle_metrics.Metric.TagsEntry
@@ -218,32 +212,6 @@ func init() { file_protos_vehicle_metric_proto_init() }
 func file_protos_vehicle_metric_proto_init() {
 	if File_protos_vehicle_metric_proto != nil {
 		return
-	}
-	if !protoimpl.UnsafeEnabled {
-		file_protos_vehicle_metric_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*VehicleMetrics); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_protos_vehicle_metric_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Metric); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
