@@ -9,7 +9,7 @@ require 'google/protobuf/timestamp_pb'
 
 descriptor_data = "\n\x12vehicle_data.proto\x12\x16telemetry.vehicle_data\x1a\x1fgoogle/protobuf/timestamp.proto\"4\n\rLocationValue\x12\x10\n\x08latitude\x18\x01 \x01(\x01\x12\x11\n\tlongitude\x18\x02 \x01(\x01\"\x86\x01\n\x05\x44oors\x12\x13\n\x0b\x44riverFront\x18\x01 \x01(\x08\x12\x12\n\nDriverRear\x18\x02 \x01(\x08\x12\x16\n\x0ePassengerFront\x18\x03 \x01(\x08\x12\x15\n\rPassengerRear\x18\x04 \x01(\x08\x12\x12\n\nTrunkFront\x18\x05 \x01(\x08\x12\x11\n\tTrunkRear\x18\x06 \x01(\x08\"\x9b\x02\n\x0cTireLocation\x12\x12\n\nfront_left\x18\x01 \x01(\x08\x12\x13\n\x0b\x66ront_right\x18\x02 \x01(\x08\x12\x11\n\trear_left\x18\x03 \x01(\x08\x12\x12\n\nrear_right\x18\x04 \x01(\x08\x12\x1f\n\x17semi_middle_axle_left_2\x18\x05 \x01(\x08\x12 \n\x18semi_middle_axle_right_2\x18\x06 \x01(\x08\x12\x1b\n\x13semi_rear_axle_left\x18\x07 \x01(\x08\x12\x1c\n\x14semi_rear_axle_right\x18\x08 \x01(\x08\x12\x1d\n\x15semi_rear_axle_left_2\x18\t \x01(\x08\x12\x1e\n\x16semi_rear_axle_right_2\x18\n \x01(\x08\"4\n\x04Time\x12\x0c\n\x04hour\x18\x01 \x01(\x05\x12\x0e\n\x06minute\x18\x02 \x01(\x05\x12\x0e\n\x06second\x18\x03 \x01(\x05\"\xde\x1c\n\x05Value\x12\x16\n\x0cstring_value\x18\x01 \x01(\tH\x00\x12\x13\n\tint_value\x18\x02 \x01(\x05H\x00\x12\x14\n\nlong_value\x18\x03 \x01(\x03H\x00\x12\x15\n\x0b\x66loat_value\x18\x04 \x01(\x02H\x00\x12\x16\n\x0c\x64ouble_value\x18\x05 \x01(\x01H\x00\x12\x17\n\rboolean_value\x18\x06 \x01(\x08H\x00\x12?\n\x0elocation_value\x18\x07 \x01(\x0b\x32%.telemetry.vehicle_data.LocationValueH\x00\x12?\n\x0e\x63harging_value\x18\x08 \x01(\x0e\x32%.telemetry.vehicle_data.ChargingStateH\x00\x12?\n\x11shift_state_value\x18\t \x01(\x0e\x32\".telemetry.vehicle_data.ShiftStateH\x00\x12\x11\n\x07invalid\x18\n \x01(\x08H\x00\x12J\n\x17lane_assist_level_value\x18\x0b \x01(\x0e\x32\'.telemetry.vehicle_data.LaneAssistLevelH\x00\x12[\n\x1dscheduled_charging_mode_value\x18\x0c \x01(\x0e\x32\x32.telemetry.vehicle_data.ScheduledChargingModeValueH\x00\x12J\n\x17sentry_mode_state_value\x18\r \x01(\x0e\x32\'.telemetry.vehicle_data.SentryModeStateH\x00\x12L\n\x18speed_assist_level_value\x18\x0e \x01(\x0e\x32(.telemetry.vehicle_data.SpeedAssistLevelH\x00\x12@\n\x0f\x62ms_state_value\x18\x0f \x01(\x0e\x32%.telemetry.vehicle_data.BMSStateValueH\x00\x12\x43\n\x13\x62uckle_status_value\x18\x10 \x01(\x0e\x32$.telemetry.vehicle_data.BuckleStatusH\x00\x12>\n\x0e\x63\x61r_type_value\x18\x11 \x01(\x0e\x32$.telemetry.vehicle_data.CarTypeValueH\x00\x12\x44\n\x11\x63harge_port_value\x18\x12 \x01(\x0e\x32\'.telemetry.vehicle_data.ChargePortValueH\x00\x12O\n\x17\x63harge_port_latch_value\x18\x13 \x01(\x0e\x32,.telemetry.vehicle_data.ChargePortLatchValueH\x00\x12\x33\n\ndoor_value\x18\x15 \x01(\x0b\x32\x1d.telemetry.vehicle_data.DoorsH\x00\x12P\n\x1a\x64rive_inverter_state_value\x18\x16 \x01(\x0e\x32*.telemetry.vehicle_data.DriveInverterStateH\x00\x12?\n\x11hvil_status_value\x18\x17 \x01(\x0e\x32\".telemetry.vehicle_data.HvilStatusH\x00\x12\x41\n\x12window_state_value\x18\x18 \x01(\x0e\x32#.telemetry.vehicle_data.WindowStateH\x00\x12L\n\x18seat_fold_position_value\x18\x19 \x01(\x0e\x32(.telemetry.vehicle_data.SeatFoldPositionH\x00\x12L\n\x18tractor_air_status_value\x18\x1a \x01(\x0e\x32(.telemetry.vehicle_data.TractorAirStatusH\x00\x12G\n\x15\x66ollow_distance_value\x18\x1b \x01(\x0e\x32&.telemetry.vehicle_data.FollowDistanceH\x00\x12\x62\n#forward_collision_sensitivity_value\x18\x1c \x01(\x0e\x32\x33.telemetry.vehicle_data.ForwardCollisionSensitivityH\x00\x12W\n\x1eguest_mode_mobile_access_value\x18\x1d \x01(\x0e\x32-.telemetry.vehicle_data.GuestModeMobileAccessH\x00\x12L\n\x18trailer_air_status_value\x18\x1e \x01(\x0e\x32(.telemetry.vehicle_data.TrailerAirStatusH\x00\x12\x32\n\ntime_value\x18\x1f \x01(\x0b\x32\x1c.telemetry.vehicle_data.TimeH\x00\x12W\n\x1b\x64\x65tailed_charge_state_value\x18  \x01(\x0e\x32\x30.telemetry.vehicle_data.DetailedChargeStateValueH\x00\x12I\n\x14hvac_auto_mode_value\x18! \x01(\x0e\x32).telemetry.vehicle_data.HvacAutoModeStateH\x00\x12h\n$cabin_overheat_protection_mode_value\x18\" \x01(\x0e\x32\x38.telemetry.vehicle_data.CabinOverheatProtectionModeStateH\x00\x12w\n1cabin_overheat_protection_temperature_limit_value\x18# \x01(\x0e\x32:.telemetry.vehicle_data.ClimateOverheatProtectionTempLimitH\x00\x12\x46\n\x12\x64\x65\x66rost_mode_value\x18$ \x01(\x0e\x32(.telemetry.vehicle_data.DefrostModeStateH\x00\x12S\n\x19\x63limate_keeper_mode_value\x18% \x01(\x0e\x32..telemetry.vehicle_data.ClimateKeeperModeStateH\x00\x12\x42\n\x10hvac_power_value\x18& \x01(\x0e\x32&.telemetry.vehicle_data.HvacPowerStateH\x00\x12\x43\n\x13tire_location_value\x18\' \x01(\x0b\x32$.telemetry.vehicle_data.TireLocationH\x00\x12\x41\n\x12\x66\x61st_charger_value\x18( \x01(\x0e\x32#.telemetry.vehicle_data.FastChargerH\x00\x12=\n\x10\x63\x61\x62le_type_value\x18) \x01(\x0e\x32!.telemetry.vehicle_data.CableTypeH\x00\x12O\n\x17tonneau_tent_mode_value\x18* \x01(\x0e\x32,.telemetry.vehicle_data.TonneauTentModeStateH\x00\x12N\n\x16tonneau_position_value\x18+ \x01(\x0e\x32,.telemetry.vehicle_data.TonneauPositionStateH\x00\x12M\n\x15powershare_type_value\x18, \x01(\x0e\x32,.telemetry.vehicle_data.PowershareTypeStatusH\x00\x12I\n\x16powershare_state_value\x18- \x01(\x0e\x32\'.telemetry.vehicle_data.PowershareStateH\x00\x12Z\n\x1cpowershare_stop_reason_value\x18. \x01(\x0e\x32\x32.telemetry.vehicle_data.PowershareStopReasonStatusH\x00\x12\x43\n\x13\x64isplay_state_value\x18/ \x01(\x0e\x32$.telemetry.vehicle_data.DisplayStateH\x00\x12\x43\n\x13\x64istance_unit_value\x18\x30 \x01(\x0e\x32$.telemetry.vehicle_data.DistanceUnitH\x00\x12I\n\x16temperature_unit_value\x18\x31 \x01(\x0e\x32\'.telemetry.vehicle_data.TemperatureUnitH\x00\x12\x43\n\x13pressure_unit_value\x18\x32 \x01(\x0e\x32$.telemetry.vehicle_data.PressureUnitH\x00\x12T\n\x1c\x63harge_unit_preference_value\x18\x33 \x01(\x0e\x32,.telemetry.vehicle_data.ChargeUnitPreferenceH\x00\x12J\n\x17turn_signal_state_value\x18\x34 \x01(\x0e\x32\'.telemetry.vehicle_data.TurnSignalStateH\x00\x12\x41\n\x12media_status_value\x18\x35 \x01(\x0e\x32#.telemetry.vehicle_data.MediaStatusH\x00\x12V\n\x1dsunroof_installed_state_value\x18\x36 \x01(\x0e\x32-.telemetry.vehicle_data.SunroofInstalledStateH\x00\x42\x07\n\x05value\"a\n\x05\x44\x61tum\x12*\n\x03key\x18\x01 \x01(\x0e\x32\x1d.telemetry.vehicle_data.Field\x12,\n\x05value\x18\x02 \x01(\x0b\x32\x1d.telemetry.vehicle_data.Value\"\x86\x01\n\x07Payload\x12+\n\x04\x64\x61ta\x18\x01 \x03(\x0b\x32\x1d.telemetry.vehicle_data.Datum\x12.\n\ncreated_at\x18\x02 \x01(\x0b\x32\x1a.google.protobuf.Timestamp\x12\x0b\n\x03vin\x18\x03 \x01(\t\x12\x11\n\tis_resend\x18\x04 \x01(\x08*\xae/\n\x05\x46ield\x12\x0b\n\x07Unknown\x10\x00\x12\r\n\tDriveRail\x10\x01\x12\x0f\n\x0b\x43hargeState\x10\x02\x12\x19\n\x15\x42msFullchargecomplete\x10\x03\x12\x10\n\x0cVehicleSpeed\x10\x04\x12\x0c\n\x08Odometer\x10\x05\x12\x0f\n\x0bPackVoltage\x10\x06\x12\x0f\n\x0bPackCurrent\x10\x07\x12\x07\n\x03Soc\x10\x08\x12\x0e\n\nDCDCEnable\x10\t\x12\x08\n\x04Gear\x10\n\x12\x17\n\x13IsolationResistance\x10\x0b\x12\x11\n\rPedalPosition\x10\x0c\x12\x0e\n\nBrakePedal\x10\r\x12\x0c\n\x08\x44iStateR\x10\x0e\x12\x10\n\x0c\x44iHeatsinkTR\x10\x0f\x12\x10\n\x0c\x44iAxleSpeedR\x10\x10\x12\x11\n\rDiTorquemotor\x10\x11\x12\x11\n\rDiStatorTempR\x10\x12\x12\x0b\n\x07\x44iVBatR\x10\x13\x12\x13\n\x0f\x44iMotorCurrentR\x10\x14\x12\x0c\n\x08Location\x10\x15\x12\x0c\n\x08GpsState\x10\x16\x12\x0e\n\nGpsHeading\x10\x17\x12\x16\n\x12NumBrickVoltageMax\x10\x18\x12\x13\n\x0f\x42rickVoltageMax\x10\x19\x12\x16\n\x12NumBrickVoltageMin\x10\x1a\x12\x13\n\x0f\x42rickVoltageMin\x10\x1b\x12\x14\n\x10NumModuleTempMax\x10\x1c\x12\x11\n\rModuleTempMax\x10\x1d\x12\x14\n\x10NumModuleTempMin\x10\x1e\x12\x11\n\rModuleTempMin\x10\x1f\x12\x0e\n\nRatedRange\x10 \x12\x08\n\x04Hvil\x10!\x12\x16\n\x12\x44\x43\x43hargingEnergyIn\x10\"\x12\x13\n\x0f\x44\x43\x43hargingPower\x10#\x12\x16\n\x12\x41\x43\x43hargingEnergyIn\x10$\x12\x13\n\x0f\x41\x43\x43hargingPower\x10%\x12\x12\n\x0e\x43hargeLimitSoc\x10&\x12\x16\n\x12\x46\x61stChargerPresent\x10\'\x12\x13\n\x0f\x45stBatteryRange\x10(\x12\x15\n\x11IdealBatteryRange\x10)\x12\x10\n\x0c\x42\x61tteryLevel\x10*\x12\x14\n\x10TimeToFullCharge\x10+\x12\x1e\n\x1aScheduledChargingStartTime\x10,\x12\x1c\n\x18ScheduledChargingPending\x10-\x12\x1a\n\x16ScheduledDepartureTime\x10.\x12\x1a\n\x16PreconditioningEnabled\x10/\x12\x19\n\x15ScheduledChargingMode\x10\x30\x12\x0e\n\nChargeAmps\x10\x31\x12\x17\n\x13\x43hargeEnableRequest\x10\x32\x12\x11\n\rChargerPhases\x10\x33\x12\x1d\n\x19\x43hargePortColdWeatherMode\x10\x34\x12\x18\n\x14\x43hargeCurrentRequest\x10\x35\x12\x1b\n\x17\x43hargeCurrentRequestMax\x10\x36\x12\x13\n\x0f\x42\x61tteryHeaterOn\x10\x37\x12\x18\n\x14NotEnoughPowerToHeat\x10\x38\x12\"\n\x1eSuperchargerSessionTripPlanner\x10\x39\x12\r\n\tDoorState\x10:\x12\n\n\x06Locked\x10;\x12\x0c\n\x08\x46\x64Window\x10<\x12\x0c\n\x08\x46pWindow\x10=\x12\x0c\n\x08RdWindow\x10>\x12\x0c\n\x08RpWindow\x10?\x12\x0f\n\x0bVehicleName\x10@\x12\x0e\n\nSentryMode\x10\x41\x12\x12\n\x0eSpeedLimitMode\x10\x42\x12\x13\n\x0f\x43urrentLimitMph\x10\x43\x12\x0b\n\x07Version\x10\x44\x12\x12\n\x0eTpmsPressureFl\x10\x45\x12\x12\n\x0eTpmsPressureFr\x10\x46\x12\x12\n\x0eTpmsPressureRl\x10G\x12\x12\n\x0eTpmsPressureRr\x10H\x12\x1e\n\x1aSemitruckTpmsPressureRe1L0\x10I\x12\x1e\n\x1aSemitruckTpmsPressureRe1L1\x10J\x12\x1e\n\x1aSemitruckTpmsPressureRe1R0\x10K\x12\x1e\n\x1aSemitruckTpmsPressureRe1R1\x10L\x12\x1e\n\x1aSemitruckTpmsPressureRe2L0\x10M\x12\x1e\n\x1aSemitruckTpmsPressureRe2L1\x10N\x12\x1e\n\x1aSemitruckTpmsPressureRe2R0\x10O\x12\x1e\n\x1aSemitruckTpmsPressureRe2R1\x10P\x12\x1e\n\x1aTpmsLastSeenPressureTimeFl\x10Q\x12\x1e\n\x1aTpmsLastSeenPressureTimeFr\x10R\x12\x1e\n\x1aTpmsLastSeenPressureTimeRl\x10S\x12\x1e\n\x1aTpmsLastSeenPressureTimeRr\x10T\x12\x0e\n\nInsideTemp\x10U\x12\x0f\n\x0bOutsideTemp\x10V\x12\x12\n\x0eSeatHeaterLeft\x10W\x12\x13\n\x0fSeatHeaterRight\x10X\x12\x16\n\x12SeatHeaterRearLeft\x10Y\x12\x17\n\x13SeatHeaterRearRight\x10Z\x12\x18\n\x14SeatHeaterRearCenter\x10[\x12\x17\n\x13\x41utoSeatClimateLeft\x10\\\x12\x18\n\x14\x41utoSeatClimateRight\x10]\x12\x12\n\x0e\x44riverSeatBelt\x10^\x12\x15\n\x11PassengerSeatBelt\x10_\x12\x16\n\x12\x44riverSeatOccupied\x10`\x12&\n\"SemitruckPassengerSeatFoldPosition\x10\x61\x12\x17\n\x13LateralAcceleration\x10\x62\x12\x1c\n\x18LongitudinalAcceleration\x10\x63\x12\x10\n\x0c\x44\x65precated_2\x10\x64\x12\x12\n\x0e\x43ruiseSetSpeed\x10\x65\x12\x16\n\x12LifetimeEnergyUsed\x10\x66\x12\x1b\n\x17LifetimeEnergyUsedDrive\x10g\x12#\n\x1fSemitruckTractorParkBrakeStatus\x10h\x12#\n\x1fSemitruckTrailerParkBrakeStatus\x10i\x12\x11\n\rBrakePedalPos\x10j\x12\x14\n\x10RouteLastUpdated\x10k\x12\r\n\tRouteLine\x10l\x12\x12\n\x0eMilesToArrival\x10m\x12\x14\n\x10MinutesToArrival\x10n\x12\x12\n\x0eOriginLocation\x10o\x12\x17\n\x13\x44\x65stinationLocation\x10p\x12\x0b\n\x07\x43\x61rType\x10q\x12\x08\n\x04Trim\x10r\x12\x11\n\rExteriorColor\x10s\x12\r\n\tRoofColor\x10t\x12\x0e\n\nChargePort\x10u\x12\x13\n\x0f\x43hargePortLatch\x10v\x12\x12\n\x0e\x45xperimental_1\x10w\x12\x12\n\x0e\x45xperimental_2\x10x\x12\x12\n\x0e\x45xperimental_3\x10y\x12\x12\n\x0e\x45xperimental_4\x10z\x12\x14\n\x10GuestModeEnabled\x10{\x12\x15\n\x11PinToDriveEnabled\x10|\x12\x1e\n\x1aPairedPhoneKeyAndKeyFobQty\x10}\x12\x18\n\x14\x43ruiseFollowDistance\x10~\x12\x1c\n\x18\x41utomaticBlindSpotCamera\x10\x7f\x12#\n\x1e\x42lindSpotCollisionWarningChime\x10\x80\x01\x12\x16\n\x11SpeedLimitWarning\x10\x81\x01\x12\x1c\n\x17\x46orwardCollisionWarning\x10\x82\x01\x12\x1b\n\x16LaneDepartureAvoidance\x10\x83\x01\x12$\n\x1f\x45mergencyLaneDepartureAvoidance\x10\x84\x01\x12!\n\x1c\x41utomaticEmergencyBrakingOff\x10\x85\x01\x12\x1e\n\x19LifetimeEnergyGainedRegen\x10\x86\x01\x12\r\n\x08\x44iStateF\x10\x87\x01\x12\x0f\n\nDiStateREL\x10\x88\x01\x12\x0f\n\nDiStateRER\x10\x89\x01\x12\x11\n\x0c\x44iHeatsinkTF\x10\x8a\x01\x12\x13\n\x0e\x44iHeatsinkTREL\x10\x8b\x01\x12\x13\n\x0e\x44iHeatsinkTRER\x10\x8c\x01\x12\x11\n\x0c\x44iAxleSpeedF\x10\x8d\x01\x12\x13\n\x0e\x44iAxleSpeedREL\x10\x8e\x01\x12\x13\n\x0e\x44iAxleSpeedRER\x10\x8f\x01\x12\x15\n\x10\x44iSlaveTorqueCmd\x10\x90\x01\x12\x14\n\x0f\x44iTorqueActualR\x10\x91\x01\x12\x14\n\x0f\x44iTorqueActualF\x10\x92\x01\x12\x16\n\x11\x44iTorqueActualREL\x10\x93\x01\x12\x16\n\x11\x44iTorqueActualRER\x10\x94\x01\x12\x12\n\rDiStatorTempF\x10\x95\x01\x12\x14\n\x0f\x44iStatorTempREL\x10\x96\x01\x12\x14\n\x0f\x44iStatorTempRER\x10\x97\x01\x12\x0c\n\x07\x44iVBatF\x10\x98\x01\x12\x0e\n\tDiVBatREL\x10\x99\x01\x12\x0e\n\tDiVBatRER\x10\x9a\x01\x12\x14\n\x0f\x44iMotorCurrentF\x10\x9b\x01\x12\x16\n\x11\x44iMotorCurrentREL\x10\x9c\x01\x12\x16\n\x11\x44iMotorCurrentRER\x10\x9d\x01\x12\x14\n\x0f\x45nergyRemaining\x10\x9e\x01\x12\x10\n\x0bServiceMode\x10\x9f\x01\x12\r\n\x08\x42MSState\x10\xa0\x01\x12\x1f\n\x1aGuestModeMobileAccessState\x10\xa1\x01\x12\x11\n\x0c\x44\x65precated_1\x10\xa2\x01\x12\x14\n\x0f\x44\x65stinationName\x10\xa3\x01\x12\x11\n\x0c\x44iInverterTR\x10\xa4\x01\x12\x11\n\x0c\x44iInverterTF\x10\xa5\x01\x12\x13\n\x0e\x44iInverterTREL\x10\xa6\x01\x12\x13\n\x0e\x44iInverterTRER\x10\xa7\x01\x12\x13\n\x0e\x45xperimental_5\x10\xa8\x01\x12\x13\n\x0e\x45xperimental_6\x10\xa9\x01\x12\x13\n\x0e\x45xperimental_7\x10\xaa\x01\x12\x13\n\x0e\x45xperimental_8\x10\xab\x01\x12\x13\n\x0e\x45xperimental_9\x10\xac\x01\x12\x14\n\x0f\x45xperimental_10\x10\xad\x01\x12\x14\n\x0f\x45xperimental_11\x10\xae\x01\x12\x14\n\x0f\x45xperimental_12\x10\xaf\x01\x12\x14\n\x0f\x45xperimental_13\x10\xb0\x01\x12\x14\n\x0f\x45xperimental_14\x10\xb1\x01\x12\x14\n\x0f\x45xperimental_15\x10\xb2\x01\x12\x18\n\x13\x44\x65tailedChargeState\x10\xb3\x01\x12 \n\x1b\x43\x61\x62inOverheatProtectionMode\x10\xb4\x01\x12,\n\'CabinOverheatProtectionTemperatureLimit\x10\xb5\x01\x12\x12\n\rCenterDisplay\x10\xb6\x01\x12\x17\n\x12\x43hargePortDoorOpen\x10\xb7\x01\x12\x13\n\x0e\x43hargerVoltage\x10\xb8\x01\x12\x16\n\x11\x43hargingCableType\x10\xb9\x01\x12\x16\n\x11\x43limateKeeperMode\x10\xba\x01\x12\x1e\n\x19\x44\x65\x66rostForPreconditioning\x10\xbb\x01\x12\x10\n\x0b\x44\x65\x66rostMode\x10\xbc\x01\x12\x16\n\x11\x45\x66\x66iciencyPackage\x10\xbd\x01\x12&\n!EstimatedHoursToChargeTermination\x10\xbe\x01\x12\x12\n\rEuropeVehicle\x10\xbf\x01\x12\'\n\"ExpectedEnergyPercentAtTripArrival\x10\xc0\x01\x12\x14\n\x0f\x46\x61stChargerType\x10\xc1\x01\x12\x18\n\x13HomelinkDeviceCount\x10\xc2\x01\x12\x13\n\x0eHomelinkNearby\x10\xc3\x01\x12\x12\n\rHvacACEnabled\x10\xc4\x01\x12\x11\n\x0cHvacAutoMode\x10\xc5\x01\x12\x11\n\x0cHvacFanSpeed\x10\xc6\x01\x12\x12\n\rHvacFanStatus\x10\xc7\x01\x12\x1f\n\x1aHvacLeftTemperatureRequest\x10\xc8\x01\x12\x0e\n\tHvacPower\x10\xc9\x01\x12 \n\x1bHvacRightTemperatureRequest\x10\xca\x01\x12\x1e\n\x19HvacSteeringWheelHeatAuto\x10\xcb\x01\x12\x1f\n\x1aHvacSteeringWheelHeatLevel\x10\xcc\x01\x12\x1b\n\x16OffroadLightbarPresent\x10\xcd\x01\x12\x18\n\x13PowershareHoursLeft\x10\xce\x01\x12#\n\x1ePowershareInstantaneousPowerKW\x10\xcf\x01\x12\x15\n\x10PowershareStatus\x10\xd0\x01\x12\x19\n\x14PowershareStopReason\x10\xd1\x01\x12\x13\n\x0ePowershareType\x10\xd2\x01\x12\x1b\n\x16RearDisplayHvacEnabled\x10\xd3\x01\x12\x14\n\x0fRearSeatHeaters\x10\xd4\x01\x12\x17\n\x12RemoteStartEnabled\x10\xd5\x01\x12\x13\n\x0eRightHandDrive\x10\xd6\x01\x12\x1d\n\x18RouteTrafficMinutesDelay\x10\xd7\x01\x12*\n%SoftwareUpdateDownloadPercentComplete\x10\xd8\x01\x12*\n%SoftwareUpdateExpectedDurationMinutes\x10\xd9\x01\x12.\n)SoftwareUpdateInstallationPercentComplete\x10\xda\x01\x12%\n SoftwareUpdateScheduledStartTime\x10\xdb\x01\x12\x1a\n\x15SoftwareUpdateVersion\x10\xdc\x01\x12\x17\n\x12TonneauOpenPercent\x10\xdd\x01\x12\x14\n\x0fTonneauPosition\x10\xde\x01\x12\x14\n\x0fTonneauTentMode\x10\xdf\x01\x12\x15\n\x10TpmsHardWarnings\x10\xe0\x01\x12\x15\n\x10TpmsSoftWarnings\x10\xe1\x01\x12\x15\n\x10ValetModeEnabled\x10\xe2\x01\x12\x0e\n\tWheelType\x10\xe3\x01\x12\x15\n\x10WiperHeatEnabled\x10\xe4\x01\x12\x12\n\rLocatedAtHome\x10\xe5\x01\x12\x12\n\rLocatedAtWork\x10\xe6\x01\x12\x16\n\x11LocatedAtFavorite\x10\xe7\x01\x12\x18\n\x13SettingDistanceUnit\x10\xe8\x01\x12\x1b\n\x16SettingTemperatureUnit\x10\xe9\x01\x12\x16\n\x11Setting24HourTime\x10\xea\x01\x12\x1c\n\x17SettingTirePressureUnit\x10\xeb\x01\x12\x16\n\x11SettingChargeUnit\x10\xec\x01\x12 \n\x1b\x43limateSeatCoolingFrontLeft\x10\xed\x01\x12!\n\x1c\x43limateSeatCoolingFrontRight\x10\xee\x01\x12\x18\n\x13LightsHazardsActive\x10\xef\x01\x12\x15\n\x10LightsTurnSignal\x10\xf0\x01\x12\x14\n\x0fLightsHighBeams\x10\xf1\x01\x12\x18\n\x13MediaPlaybackStatus\x10\xf2\x01\x12\x18\n\x13MediaPlaybackSource\x10\xf3\x01\x12\x15\n\x10MediaAudioVolume\x10\xf4\x01\x12\x1c\n\x17MediaNowPlayingDuration\x10\xf5\x01\x12\x1b\n\x16MediaNowPlayingElapsed\x10\xf6\x01\x12\x1a\n\x15MediaNowPlayingArtist\x10\xf7\x01\x12\x19\n\x14MediaNowPlayingTitle\x10\xf8\x01\x12\x19\n\x14MediaNowPlayingAlbum\x10\xf9\x01\x12\x1b\n\x16MediaNowPlayingStation\x10\xfa\x01\x12\x1e\n\x19MediaAudioVolumeIncrement\x10\xfb\x01\x12\x18\n\x13MediaAudioVolumeMax\x10\xfc\x01\x12\x15\n\x10SunroofInstalled\x10\xfd\x01\x12\x14\n\x0fSeatVentEnabled\x10\xfe\x01\x12\x17\n\x12RearDefrostEnabled\x10\xff\x01\x12\x1a\n\x15\x43hargeRateMilePerHour\x10\x80\x02\x12\x11\n\x0c\x44\x65precated_3\x10\x81\x02*\xbf\x01\n\rChargingState\x12\x16\n\x12\x43hargeStateUnknown\x10\x00\x12\x1b\n\x17\x43hargeStateDisconnected\x10\x01\x12\x16\n\x12\x43hargeStateNoPower\x10\x02\x12\x17\n\x13\x43hargeStateStarting\x10\x03\x12\x17\n\x13\x43hargeStateCharging\x10\x04\x12\x17\n\x13\x43hargeStateComplete\x10\x05\x12\x16\n\x12\x43hargeStateStopped\x10\x06*\x82\x02\n\x18\x44\x65tailedChargeStateValue\x12\x1e\n\x1a\x44\x65tailedChargeStateUnknown\x10\x00\x12#\n\x1f\x44\x65tailedChargeStateDisconnected\x10\x01\x12\x1e\n\x1a\x44\x65tailedChargeStateNoPower\x10\x02\x12\x1f\n\x1b\x44\x65tailedChargeStateStarting\x10\x03\x12\x1f\n\x1b\x44\x65tailedChargeStateCharging\x10\x04\x12\x1f\n\x1b\x44\x65tailedChargeStateComplete\x10\x05\x12\x1e\n\x1a\x44\x65tailedChargeStateStopped\x10\x06*\x91\x01\n\nShiftState\x12\x15\n\x11ShiftStateUnknown\x10\x00\x12\x15\n\x11ShiftStateInvalid\x10\x01\x12\x0f\n\x0bShiftStateP\x10\x02\x12\x0f\n\x0bShiftStateR\x10\x03\x12\x0f\n\x0bShiftStateN\x10\x04\x12\x0f\n\x0bShiftStateD\x10\x05\x12\x11\n\rShiftStateSNA\x10\x06*\xbe\x01\n\x0e\x46ollowDistance\x12\x19\n\x15\x46ollowDistanceUnknown\x10\x00\x12\x13\n\x0f\x46ollowDistance1\x10\x01\x12\x13\n\x0f\x46ollowDistance2\x10\x02\x12\x13\n\x0f\x46ollowDistance3\x10\x03\x12\x13\n\x0f\x46ollowDistance4\x10\x04\x12\x13\n\x0f\x46ollowDistance5\x10\x05\x12\x13\n\x0f\x46ollowDistance6\x10\x06\x12\x13\n\x0f\x46ollowDistance7\x10\x07*\xdc\x01\n\x1b\x46orwardCollisionSensitivity\x12&\n\"ForwardCollisionSensitivityUnknown\x10\x00\x12\"\n\x1e\x46orwardCollisionSensitivityOff\x10\x01\x12#\n\x1f\x46orwardCollisionSensitivityLate\x10\x02\x12&\n\"ForwardCollisionSensitivityAverage\x10\x03\x12$\n ForwardCollisionSensitivityEarly\x10\x04*\xe4\x06\n\x15GuestModeMobileAccess\x12 \n\x1cGuestModeMobileAccessUnknown\x10\x00\x12\x1d\n\x19GuestModeMobileAccessInit\x10\x01\x12)\n%GuestModeMobileAccessNotAuthenticated\x10\x02\x12&\n\"GuestModeMobileAccessAuthenticated\x10\x03\x12\'\n#GuestModeMobileAccessAbortedDriving\x10\x04\x12\x30\n,GuestModeMobileAccessAbortedUsingRemoteStart\x10\x05\x12,\n(GuestModeMobileAccessAbortedUsingBLEKeys\x10\x06\x12)\n%GuestModeMobileAccessAbortedValetMode\x10\x07\x12,\n(GuestModeMobileAccessAbortedGuestModeOff\x10\x08\x12\x35\n1GuestModeMobileAccessAbortedDriveAuthTimeExceeded\x10\t\x12.\n*GuestModeMobileAccessAbortedNoDataReceived\x10\n\x12\x31\n-GuestModeMobileAccessRequestingFromMothership\x10\x0b\x12,\n(GuestModeMobileAccessRequestingFromAuthD\x10\x0c\x12+\n\'GuestModeMobileAccessAbortedFetchFailed\x10\r\x12/\n+GuestModeMobileAccessAbortedBadDataReceived\x10\x0e\x12&\n\"GuestModeMobileAccessShowingQRCode\x10\x0f\x12#\n\x1fGuestModeMobileAccessSwipedAway\x10\x10\x12/\n+GuestModeMobileAccessDismissedQRCodeExpired\x10\x11\x12\x31\n-GuestModeMobileAccessSucceededPairedNewBLEKey\x10\x12*}\n\x0fLaneAssistLevel\x12\x1a\n\x16LaneAssistLevelUnknown\x10\x00\x12\x17\n\x13LaneAssistLevelNone\x10\x01\x12\x1a\n\x16LaneAssistLevelWarning\x10\x02\x12\x19\n\x15LaneAssistLevelAssist\x10\x03*\xa1\x01\n\x1aScheduledChargingModeValue\x12 \n\x1cScheduledChargingModeUnknown\x10\x00\x12\x1c\n\x18ScheduledChargingModeOff\x10\x01\x12 \n\x1cScheduledChargingModeStartAt\x10\x02\x12!\n\x1dScheduledChargingModeDepartBy\x10\x03*\xc6\x01\n\x0fSentryModeState\x12\x1a\n\x16SentryModeStateUnknown\x10\x00\x12\x16\n\x12SentryModeStateOff\x10\x01\x12\x17\n\x13SentryModeStateIdle\x10\x02\x12\x18\n\x14SentryModeStateArmed\x10\x03\x12\x18\n\x14SentryModeStateAware\x10\x04\x12\x18\n\x14SentryModeStatePanic\x10\x05\x12\x18\n\x14SentryModeStateQuiet\x10\x06*\x81\x01\n\x10SpeedAssistLevel\x12\x1b\n\x17SpeedAssistLevelUnknown\x10\x00\x12\x18\n\x14SpeedAssistLevelNone\x10\x01\x12\x1b\n\x17SpeedAssistLevelDisplay\x10\x02\x12\x19\n\x15SpeedAssistLevelChime\x10\x03*\xe7\x01\n\rBMSStateValue\x12\x13\n\x0f\x42MSStateUnknown\x10\x00\x12\x13\n\x0f\x42MSStateStandby\x10\x01\x12\x11\n\rBMSStateDrive\x10\x02\x12\x13\n\x0f\x42MSStateSupport\x10\x03\x12\x12\n\x0e\x42MSStateCharge\x10\x04\x12\x10\n\x0c\x42MSStateFEIM\x10\x05\x12\x16\n\x12\x42MSStateClearFault\x10\x06\x12\x11\n\rBMSStateFault\x10\x07\x12\x10\n\x0c\x42MSStateWeld\x10\x08\x12\x10\n\x0c\x42MSStateTest\x10\t\x12\x0f\n\x0b\x42MSStateSNA\x10\n*t\n\x0c\x42uckleStatus\x12\x17\n\x13\x42uckleStatusUnknown\x10\x00\x12\x19\n\x15\x42uckleStatusUnlatched\x10\x01\x12\x17\n\x13\x42uckleStatusLatched\x10\x02\x12\x17\n\x13\x42uckleStatusFaulted\x10\x03*\x9b\x01\n\x0c\x43\x61rTypeValue\x12\x12\n\x0e\x43\x61rTypeUnknown\x10\x00\x12\x11\n\rCarTypeModelS\x10\x01\x12\x11\n\rCarTypeModelX\x10\x02\x12\x11\n\rCarTypeModel3\x10\x03\x12\x11\n\rCarTypeModelY\x10\x04\x12\x14\n\x10\x43\x61rTypeSemiTruck\x10\x05\x12\x15\n\x11\x43\x61rTypeCybertruck\x10\x06*q\n\x0f\x43hargePortValue\x12\x15\n\x11\x43hargePortUnknown\x10\x00\x12\x10\n\x0c\x43hargePortUS\x10\x01\x12\x10\n\x0c\x43hargePortEU\x10\x02\x12\x10\n\x0c\x43hargePortGB\x10\x03\x12\x11\n\rChargePortCCS\x10\x04*\xa2\x01\n\x14\x43hargePortLatchValue\x12\x1a\n\x16\x43hargePortLatchUnknown\x10\x00\x12\x16\n\x12\x43hargePortLatchSNA\x10\x01\x12\x1d\n\x19\x43hargePortLatchDisengaged\x10\x02\x12\x1a\n\x16\x43hargePortLatchEngaged\x10\x03\x12\x1b\n\x17\x43hargePortLatchBlocking\x10\x04*\xcd\x01\n\x12\x44riveInverterState\x12\x1d\n\x19\x44riveInverterStateUnknown\x10\x00\x12!\n\x1d\x44riveInverterStateUnavailable\x10\x01\x12\x1d\n\x19\x44riveInverterStateStandby\x10\x02\x12\x1b\n\x17\x44riveInverterStateFault\x10\x03\x12\x1b\n\x17\x44riveInverterStateAbort\x10\x04\x12\x1c\n\x18\x44riveInverterStateEnable\x10\x05*J\n\nHvilStatus\x12\x15\n\x11HvilStatusUnknown\x10\x00\x12\x13\n\x0fHvilStatusFault\x10\x01\x12\x10\n\x0cHvilStatusOK\x10\x02*q\n\x0bWindowState\x12\x16\n\x12WindowStateUnknown\x10\x00\x12\x15\n\x11WindowStateClosed\x10\x01\x12\x1c\n\x18WindowStatePartiallyOpen\x10\x02\x12\x15\n\x11WindowStateOpened\x10\x03*\xc2\x01\n\x10SeatFoldPosition\x12\x1b\n\x17SeatFoldPositionUnknown\x10\x00\x12\x17\n\x13SeatFoldPositionSNA\x10\x01\x12\x1b\n\x17SeatFoldPositionFaulted\x10\x02\x12!\n\x1dSeatFoldPositionNotConfigured\x10\x03\x12\x1a\n\x16SeatFoldPositionFolded\x10\x04\x12\x1c\n\x18SeatFoldPositionUnfolded\x10\x05*\x8e\x02\n\x10TractorAirStatus\x12\x1b\n\x17TractorAirStatusUnknown\x10\x00\x12 \n\x1cTractorAirStatusNotAvailable\x10\x01\x12\x19\n\x15TractorAirStatusError\x10\x02\x12\x1b\n\x17TractorAirStatusCharged\x10\x03\x12\x30\n,TractorAirStatusBuildingPressureIntermediate\x10\x04\x12\x32\n.TractorAirStatusExhaustingPressureIntermediate\x10\x05\x12\x1d\n\x19TractorAirStatusExhausted\x10\x06*\xa8\x02\n\x10TrailerAirStatus\x12\x1b\n\x17TrailerAirStatusUnknown\x10\x00\x12\x17\n\x13TrailerAirStatusSNA\x10\x01\x12\x1b\n\x17TrailerAirStatusInvalid\x10\x02\x12\x1f\n\x1bTrailerAirStatusBobtailMode\x10\x03\x12\x1b\n\x17TrailerAirStatusCharged\x10\x04\x12\x30\n,TrailerAirStatusBuildingPressureIntermediate\x10\x05\x12\x32\n.TrailerAirStatusExhaustingPressureIntermediate\x10\x06\x12\x1d\n\x19TrailerAirStatusExhausted\x10\x07*i\n\x11HvacAutoModeState\x12\x1c\n\x18HvacAutoModeStateUnknown\x10\x00\x12\x17\n\x13HvacAutoModeStateOn\x10\x01\x12\x1d\n\x19HvacAutoModeStateOverride\x10\x02*\xcd\x01\n CabinOverheatProtectionModeState\x12+\n\'CabinOverheatProtectionModeStateUnknown\x10\x00\x12\'\n#CabinOverheatProtectionModeStateOff\x10\x01\x12&\n\"CabinOverheatProtectionModeStateOn\x10\x02\x12+\n\'CabinOverheatProtectionModeStateFanOnly\x10\x03*\xd8\x01\n\"ClimateOverheatProtectionTempLimit\x12-\n)ClimateOverheatProtectionTempLimitUnknown\x10\x00\x12*\n&ClimateOverheatProtectionTempLimitHigh\x10\x01\x12,\n(ClimateOverheatProtectionTempLimitMedium\x10\x02\x12)\n%ClimateOverheatProtectionTempLimitLow\x10\x03*\x9c\x01\n\x10\x44\x65\x66rostModeState\x12\x1b\n\x17\x44\x65\x66rostModeStateUnknown\x10\x00\x12\x17\n\x13\x44\x65\x66rostModeStateOff\x10\x01\x12\x1a\n\x16\x44\x65\x66rostModeStateNormal\x10\x02\x12\x17\n\x13\x44\x65\x66rostModeStateMax\x10\x03\x12\x1d\n\x19\x44\x65\x66rostModeStateAutoDefog\x10\x04*\xb8\x01\n\x16\x43limateKeeperModeState\x12!\n\x1d\x43limateKeeperModeStateUnknown\x10\x00\x12\x1d\n\x19\x43limateKeeperModeStateOff\x10\x01\x12\x1c\n\x18\x43limateKeeperModeStateOn\x10\x02\x12\x1d\n\x19\x43limateKeeperModeStateDog\x10\x03\x12\x1f\n\x1b\x43limateKeeperModeStateParty\x10\x04*\x9b\x01\n\x0eHvacPowerState\x12\x19\n\x15HvacPowerStateUnknown\x10\x00\x12\x15\n\x11HvacPowerStateOff\x10\x01\x12\x14\n\x10HvacPowerStateOn\x10\x02\x12\x1e\n\x1aHvacPowerStatePrecondition\x10\x03\x12!\n\x1dHvacPowerStateOverheatProtect\x10\x04*\xed\x01\n\x0b\x46\x61stCharger\x12\x16\n\x12\x46\x61stChargerUnknown\x10\x00\x12\x1b\n\x17\x46\x61stChargerSupercharger\x10\x01\x12\x16\n\x12\x46\x61stChargerCHAdeMO\x10\x02\x12\x11\n\rFastChargerGB\x10\x03\x12\x1e\n\x1a\x46\x61stChargerACSingleWireCAN\x10\x04\x12\x14\n\x10\x46\x61stChargerCombo\x10\x05\x12\x1e\n\x1a\x46\x61stChargerMCSingleWireCAN\x10\x06\x12\x14\n\x10\x46\x61stChargerOther\x10\x07\x12\x12\n\x0e\x46\x61stChargerSNA\x10\x08*\x7f\n\tCableType\x12\x14\n\x10\x43\x61\x62leTypeUnknown\x10\x00\x12\x10\n\x0c\x43\x61\x62leTypeIEC\x10\x01\x12\x10\n\x0c\x43\x61\x62leTypeSAE\x10\x02\x12\x12\n\x0e\x43\x61\x62leTypeGB_AC\x10\x03\x12\x12\n\x0e\x43\x61\x62leTypeGB_DC\x10\x04\x12\x10\n\x0c\x43\x61\x62leTypeSNA\x10\x05*\xb9\x01\n\x14TonneauTentModeState\x12\x1f\n\x1bTonneauTentModeStateUnknown\x10\x00\x12 \n\x1cTonneauTentModeStateInactive\x10\x01\x12\x1e\n\x1aTonneauTentModeStateMoving\x10\x02\x12\x1e\n\x1aTonneauTentModeStateFailed\x10\x03\x12\x1e\n\x1aTonneauTentModeStateActive\x10\x04*\xc2\x01\n\x14TonneauPositionState\x12\x1f\n\x1bTonneauPositionStateUnknown\x10\x00\x12\x1f\n\x1bTonneauPositionStateInvalid\x10\x01\x12\x1e\n\x1aTonneauPositionStateClosed\x10\x02\x12%\n!TonneauPositionStatePartiallyOpen\x10\x03\x12!\n\x1dTonneauPositionStateFullyOpen\x10\x04*\xe7\x01\n\x0fPowershareState\x12\x1a\n\x16PowershareStateUnknown\x10\x00\x12\x1b\n\x17PowershareStateInactive\x10\x01\x12\x1e\n\x1aPowershareStateHandshaking\x10\x02\x12\x17\n\x13PowershareStateInit\x10\x03\x12\x1a\n\x16PowershareStateEnabled\x10\x04\x12*\n&PowershareStateEnabledReconnectingSoon\x10\x05\x12\x1a\n\x16PowershareStateStopped\x10\x06*\xd8\x02\n\x1aPowershareStopReasonStatus\x12%\n!PowershareStopReasonStatusUnknown\x10\x00\x12\"\n\x1ePowershareStopReasonStatusNone\x10\x01\x12\'\n#PowershareStopReasonStatusSOCTooLow\x10\x02\x12#\n\x1fPowershareStopReasonStatusRetry\x10\x03\x12#\n\x1fPowershareStopReasonStatusFault\x10\x04\x12\"\n\x1ePowershareStopReasonStatusUser\x10\x05\x12*\n&PowershareStopReasonStatusReconnecting\x10\x06\x12,\n(PowershareStopReasonStatusAuthentication\x10\x07*\x91\x01\n\x14PowershareTypeStatus\x12\x1f\n\x1bPowershareTypeStatusUnknown\x10\x00\x12\x1c\n\x18PowershareTypeStatusNone\x10\x01\x12\x1c\n\x18PowershareTypeStatusLoad\x10\x02\x12\x1c\n\x18PowershareTypeStatusHome\x10\x03*\x95\x02\n\x0c\x44isplayState\x12\x17\n\x13\x44isplayStateUnknown\x10\x00\x12\x13\n\x0f\x44isplayStateOff\x10\x01\x12\x13\n\x0f\x44isplayStateDim\x10\x02\x12\x19\n\x15\x44isplayStateAccessory\x10\x03\x12\x12\n\x0e\x44isplayStateOn\x10\x04\x12\x17\n\x13\x44isplayStateDriving\x10\x05\x12\x18\n\x14\x44isplayStateCharging\x10\x06\x12\x14\n\x10\x44isplayStateLock\x10\x07\x12\x16\n\x12\x44isplayStateSentry\x10\x08\x12\x13\n\x0f\x44isplayStateDog\x10\t\x12\x1d\n\x19\x44isplayStateEntertainment\x10\n*Z\n\x0c\x44istanceUnit\x12\x17\n\x13\x44istanceUnitUnknown\x10\x00\x12\x15\n\x11\x44istanceUnitMiles\x10\x01\x12\x1a\n\x16\x44istanceUnitKilometers\x10\x02*h\n\x0fTemperatureUnit\x12\x1a\n\x16TemperatureUnitUnknown\x10\x00\x12\x1d\n\x19TemperatureUnitFahrenheit\x10\x01\x12\x1a\n\x16TemperatureUnitCelsius\x10\x02*Q\n\x0cPressureUnit\x12\x17\n\x13PressureUnitUnknown\x10\x00\x12\x13\n\x0fPressureUnitPsi\x10\x01\x12\x13\n\x0fPressureUnitBar\x10\x02*\\\n\x14\x43hargeUnitPreference\x12\x15\n\x11\x43hargeUnitUnknown\x10\x00\x12\x16\n\x12\x43hargeUnitDistance\x10\x01\x12\x15\n\x11\x43hargeUnitPercent\x10\x02*\xb0\x01\n\x15SunroofInstalledState\x12 \n\x1cSunroofInstalledStateUnknown\x10\x00\x12%\n!SunroofInstalledStateNotInstalled\x10\x01\x12&\n\"SunroofInstalledStateGen1Installed\x10\x02\x12&\n\"SunroofInstalledStateGen2Installed\x10\x03*\x91\x01\n\x0fTurnSignalState\x12\x1a\n\x16TurnSignalStateUnknown\x10\x00\x12\x16\n\x12TurnSignalStateOff\x10\x01\x12\x17\n\x13TurnSignalStateLeft\x10\x02\x12\x18\n\x14TurnSignalStateRight\x10\x03\x12\x17\n\x13TurnSignalStateBoth\x10\x04*l\n\x0bMediaStatus\x12\x16\n\x12MediaStatusUnknown\x10\x00\x12\x16\n\x12MediaStatusStopped\x10\x01\x12\x16\n\x12MediaStatusPlaying\x10\x02\x12\x15\n\x11MediaStatusPaused\x10\x03\x42/Z-github.com/teslamotors/fleet-telemetry/protosb\x06proto3"
 
-pool = Google::Protobuf::DescriptorPool.generated_pool
+pool = ::Google::Protobuf::DescriptorPool.generated_pool
 pool.add_serialized_file(descriptor_data)
 
 module Telemetry
